+++ iris_grib/_save_rules.py
@@ -1102,7 +1102,7 @@ def set_fixed_surfaces(cube, grib, full3d_cube=None):
                           int(round(output_v[1])))
 
     if hybrid_factory is not None:
-        # Need to record ALL the level coefficents in a 'PV' vector.
+        # Need to record ALL the level coefficients in a 'PV' vector.
         level_delta_coord = hybrid_factory.delta
         sigma_coord = hybrid_factory.sigma
         model_levels = full3d_cube.coord('model_level_number').points
@@ -1111,7 +1111,7 @@ def set_fixed_surfaces(cube, grib, full3d_cube=None):
             msg = 'model_level_number is not an integer: dtype={}.'
             raise ValueError(msg.format(model_levels.dtype))
         if np.min(model_levels) < 1:
-            msg = 'model_level_number must be > 0: mininum value = {}.'
+            msg = 'model_level_number must be > 0: minimum value = {}.'
             raise ValueError(msg.format(np.min(model_levels)))
         # Need to save enough levels for indexes up to  [max(model_levels)]
         n_levels = np.max(model_levels)
@@ -1373,6 +1373,25 @@ def product_definition_template_1(cube, grib, full3d_cube=None):
     set_ensemble(cube, grib)
 
 
+def product_definition_template_6(cube, grib, full3d_cube=None):
+    """
+    Set keys within the provided grib message based on Product Definition
+    Template 4.6.
+
+    Template 4.6 is used to represent a percentile forecast at a point in time
+    interval.
+
+    """
+    eccodes.codes_set(grib, "productDefinitionTemplateNumber", 6)
+    product_definition_template_common(cube, grib, full3d_cube)
+    if not (cube.coords('percentile') and
+            len(cube.coord('percentile').points) == 1):
+        raise ValueError("A cube 'percentile' coordinate with one "
+                         "point is required, but not present.")
+    eccodes.codes_set(grib, "percentileValue",
+                      int(cube.coord('percentile').points[0]))
+
+
 def product_definition_template_8(cube, grib, full3d_cube=None):
     """
     Set keys within the provided grib message based on Product
@@ -1567,6 +1586,8 @@ def product_definition_section(cube, grib, full3d_cube=None):
         elif 'spatial_processing_type' in cube.attributes:
             # spatial process (template 4.15)
             product_definition_template_15(cube, grib, full3d_cube)
+        elif cube.coords('percentile'):
+            product_definition_template_6(cube, grib, full3d_cube)
         else:
             # forecast (template 4.0)
             product_definition_template_0(cube, grib, full3d_cube)